requirements.txt
Django==2.2
mysqlclient==1.4.6
Pillow>=7
django-simple-captcha
djangorestframework
testfixtures==6.14.1
xlrd
django-suit==2.0a1
wagtail==2.9.2
drf-yasg